Trivia file rename. Created Trivia object to hold trivia data and state. Added Trivia object to Discord bot
